Move build util and cli into src dir
