+++ package.json
@@ -4,11 +4,11 @@
   "description": "js utils for how we use JSON Schema",
   "main": "index.js",
   "bin": {
-    "json-schema-utils": "./cli/index.js"
+    "json-schema-utils": "./src/cli/index.js"
   },
   "scripts": {
     "test": "mocha -R spec 'tests/**/*.test.js'",
-    "lint": "eslint './*.js' 'tests/**/*.js'",
+    "lint": "eslint './*.js' './src/**/*.js' 'tests/**/*.js'",
     "ci": "npm run lint && npm test"
   },
   "keywords": [
